[refactor] Main.js
이상 수치 로그 기록 -> component/specific/abnormalLog.js와 module.css로 분리
