+++ src/pages/Main.module.css
@@ -240,62 +240,14 @@
   margin-bottom: 15%;
   margin-right: -10px;
 }
-/*날씨*/
-.weatherInfo {
-  font-size: clamp(20px, 2vw, 50px);
-  font-weight: bold;
-  margin-top: 1%;
-  background-color: #ffedc3; /* 오렌지색 배경 */
-  border-radius: 20px 0 0 20px; /* 모서리 */
-  padding: clamp(12px, 1vw, 100px); /* 박스 내부 여백 */
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.261); /* 부드러운 그림자 */
-  width: 100%; /* 박스가 부모의 너비를 차지하도록 */
-  color: black;
-  z-index: 800;
-}
-.weatherMinMax {
-  font-size: clamp(16px, 1vw, 24px);
-  text-align: right;
-  font-weight: 600;
-  color: black;
-}
-.weatherMinMax2 {
-  text-align: right;
-  font-size: clamp(9px, 1vw, 13px);
-  margin-bottom: 10px;
-  color: black;
-}
-.weatherTemperature {
-  font-size: clamp(25px, 3vw, 100px);
-  font-weight: bold;
-  text-align: right;
-}
-.weatherRain {
-  font-size: 2vw;
-  text-align: center;
-  color: black;
-}
-.weatherAdvice {
-  text-align: right;
-  white-space: pre-line;
-  font-size: clamp(12px, 1.2vw, 18px);
-  color: black;
-}
+
 .weatherlocation {
   font-size: clamp(15px, 1.6vw, 25px);
   width: clamp(20px, 2vw, 50px);
   margin-right: 0.5vw;
   margin-bottom: -4px;
 }
-.weatherloading {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.weatherImg {
-  width: clamp(70px, 7vw, 200px);
-  filter: drop-shadow(4px 4px 4px rgba(0, 0, 0, 0.2));
-}
+
 /*로그기록*/
 .sensorLogs {
   background-color: #ffd9b7; /* 오렌지색 배경 */
@@ -315,18 +267,9 @@
   color: grey;
 }
 
-.weatherInfo h3 {
-  font-size: 32px;
-  margin-bottom: 20px;
-}
 .sensorLogs h3 {
   font-size: 18px;
   margin-bottom: 10px;
-}
-
-.weatherInfo p,
-.sensorInfo p {
-  margin: 5px 0;
 }
 
 /* 신공학관 누르면 */
@@ -462,7 +405,7 @@
   top: -28%;
   left: 35%;
   transform: translateX(-50%);
-  background-color: rgba(251, 244, 228, 0.8);
+  background-color: white;
   color: black;
   padding: 10%;
   border-radius: 8px;
